LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/after-transition-push-expected.txt


FAIL scroll: after-transition should work on a push navigation assert_equals: expected 0 but got 431
